[MOV] rating: extract mixin in new file
This commit simply move code for both rating.mixin
and rating.parent.mixin in their own file to ease
the reading of code, and fit the guidelines.

Task-1903565
