+++ addons/rating/models/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
 from . import rating
+from . import rating_mixin
 from . import mail_thread
 from . import mail_message
